# Speed up system tests by caching yaml output