--- a/libbeat/tests/system/beat/beat.py
+++ b/libbeat/tests/system/beat/beat.py
@@ -9,6 +9,7 @@ import signal
 import sys
 import time
 import yaml
+import hashlib
 from datetime import datetime, timedelta
 
 from .compose import ComposeMixin
@@ -18,6 +19,8 @@ BEAT_REQUIRED_FIELDS = ["@timestamp",
                         "beat.name", "beat.hostname", "beat.version"]
 
 INTEGRATION_TESTS = os.environ.get('INTEGRATION_TESTS', False)
+
+yaml_cache = {}
 
 
 class TimeoutError(Exception):
@@ -487,6 +490,8 @@ class TestCase(unittest.TestCase, ComposeMixin):
         if not os.path.isfile(fields_doc):
             fields_doc = self.beat_path + "/_meta/fields.yml"
 
+        global yaml_cache
+
         # TODO: Make fields_doc path more generic to work with beat-generator
         with open(fields_doc, "r") as f:
             path = os.path.abspath(os.path.dirname(__file__) + "../../../../_meta/fields.generated.yml")
@@ -495,9 +500,15 @@ class TestCase(unittest.TestCase, ComposeMixin):
             with open(path) as f2:
                 content = f2.read()
 
-            #content = "fields:\n"
             content += f.read()
-            doc = yaml.load(content)
+
+            hash = hashlib.md5(content).hexdigest()
+            doc = ""
+            if hash in yaml_cache:
+                doc = yaml_cache[hash]
+            else:
+                doc = yaml.safe_load(content)
+                yaml_cache[hash] = doc
 
             fields = []
             dictfields = []