`yaml.load()` took each time up to 1.5s when loading the full fields.yml. This heavily slowed down the tests as the yaml file was loaded each time and event was checked if all fields are documented. This change introduces a cache for the output from the `yaml.load` based on an md5 hash.

test_system before

```
nosetests test_system.py:Test.test_network -v
----------------------------------------------------------------------
Ran 1 test in 8.948s

OK
(python-env) ruflin:system ruflin$ nosetests test_system.py -v
Test core system output. ... ok
Test core system output. ... ok
Test cpu system output. ... ok
Test cpu_ticks configuration option. ... ok
Test system/diskio output. ... SKIP: os
Test system/diskio output on linux. ... SKIP: os
Test system/filesystem output. ... ok
Test system/fsstat output. ... ok
Test system load. ... ok
Test system memory output. ... ok
Test system/network output. ... ok
Test system/process output. ... ok
Checks that the per proc stats are found in the output and ... ok
Test system/process_summary output. ... ok

----------------------------------------------------------------------
Ran 14 tests in 396.891s
```

test_system after

```
----------------------------------------------------------------------
Ran 14 tests in 396.891s

OK (SKIP=2)
(python-env) ruflin:system ruflin$ nosetests test_system.py -v
Test core system output. ... ok
Test core system output. ... ok
Test cpu system output. ... ok
Test cpu_ticks configuration option. ... ok
Test system/diskio output. ... SKIP: os
Test system/diskio output on linux. ... SKIP: os
Test system/filesystem output. ... ok
Test system/fsstat output. ... ok
Test system load. ... ok
Test system memory output. ... ok
Test system/network output. ... ok
Test system/process output. ... ok
Checks that the per proc stats are found in the output and ... ok
Test system/process_summary output. ... ok

----------------------------------------------------------------------
Ran 14 tests in 6.515s
```